+++ addOns/automation/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Fixed
+- Address HTTP authentication failure when the realm is not configured.
 
 ## [0.40.0] - 2024-05-07
 ### Changed